add tests

+++ python/cudf/cudf/core/column/string.py
@@ -5307,8 +5307,8 @@ class StringColumn(column.ColumnBase):
     def _find_first_and_last(self, value: ScalarLike) -> Tuple[int, int]:
         found_indices = libstrings.contains_re(self, f"^{value}$")
         found_indices = libcudf.unary.cast(found_indices, dtype=np.int32)
-        first = column.as_column(found_indices).find_first_value(1)
-        last = column.as_column(found_indices).find_last_value(1)
+        first = column.as_column(found_indices).find_first_value(np.int32(1))
+        last = column.as_column(found_indices).find_last_value(np.int32(1))
         return first, last
 
     def find_first_value(
